Model corrections & created UUIDAutoField

+++ app/models/Backup.py
@@ -8,7 +8,7 @@ class Backup(models.Model):
   processed = models.BooleanField(default=False)
   dump_file_name = models.TextField(blank=True)
   dump_content_type = models.TextField(blank=True)
-  dump_file_size = models.DecimalField(blank=True)
+  dump_file_size = models.PositiveBigIntegerField(blank=True)
   dump_updated_at = models.DateTimeField(blank=True)
   created_at = models.DateTimeField(auto_now_add=True)
   updated_at = models.DateTimeField(auto_now=True)
