+++ app/models/Import.py
@@ -9,7 +9,7 @@ class Import(models.Model):
   overwrite = models.BooleanField(default=True)
   data_file_name = models.TextField(blank=True)
   data_content_type = models.TextField(blank=True)
-  data_file_size = models.DecimalField(blank=True)
+  data_file_size = models.PositiveBigIntegerField(blank=True)
   data_updated_at = models.DateTimeField(blank=True)
   created_at = models.DateTimeField(auto_now_add=True)
   updated_at = models.DateTimeField(auto_now=True)
